Move foreman UI to glue foreman
