+++ src/app/views/foreman/domains/_new.html.haml
