+++ src/app/views/foreman/domains/index.html.haml
