+++ src/app/views/foreman/subnets/index.html.haml
